refactor: split out some functions again
