+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/collection/number/LongCollectionExt.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.kotlinutils.collection
+package xyz.lbres.kotlinutils.collection.number
 
 import xyz.lbres.kotlinutils.number.isZero
 import xyz.lbres.kotlinutils.utils.simpleIf
