+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/generic/GenericExt.kt
@@ -29,3 +29,4 @@ fun <T> T?.isNotNull(): Boolean = this != null
  * @return `null` if this value is `null`, or result of executing [block] on the non-null value
  */
 fun <T, S> T?.ifNotNull(block: (T) -> S): S? = if (this == null) null else block(this)
+// TODO use this
